requirements.txt
pandas
telepot
python-dateutil
pytz
psycopg2-binary
discord-webhook
#facebook-scraper
